Modularized Index.js into controllers file.
